+++ app/user/adapter/input/routes/__init__.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from app.user.adapter.input.api.v1.user import user_router as user_v1_router
+from app.user.adapter.input.routes.user import user_router as user_v1_router
 
 router = APIRouter()
 router.include_router(user_v1_router, prefix="/api/v1/user", tags=["User"])
